Add files via upload
Image Processing to measure area of arctic sea ice from 1979-2020. The data has been downloaded from NASA's official website and shows how the ice area in arctic region has been depreciating with each decade. OpenCV is used to analyze each image, 42 in total and we can see the line plot of area in the end for it.